+++ tools/nissy_extra.h
@@ -10,5 +10,6 @@ for testing purposes only.
 
 size_t gendata_h48_derive(uint8_t, const void *, void *);
 int parse_h48_solver(const char *, uint8_t [static 1], uint8_t [static 1]);
-int64_t nissy_datainfo(uint64_t, const char *, void (*)(const char *, ...));
-int64_t nissy_derivedata(const char *, const void *, void *);
+long long int nissy_datainfo(
+    uint64_t, const char *, void (*)(const char *, ...));
+long long int nissy_derivedata(const char *, const void *, void *);
